wzrd/requirements.txt
django
pydash
psycopg2
djangorestframework